backend/requirements.txt
openai
python-dotenv
uvicorn[standard]
fastapi
pydantic
requests